+++ web/app/stores/menu/menu.reducer.js
@@ -99,13 +99,17 @@ let menuData = immutable.fromJS({
   }, {
     display: 'Showcase',
     items: [{
+      id: 'showcase-algorithms',
+      display: 'Algorithms',
+      to: '/showcase/algorithms'
+    }, {
       id: 'showcase-dynamically-loading-components',
       display: 'Dynamically Loading Components',
       to: '/showcase/dynamically-loading-components'
     }, {
-      id: 'showcase-algorithms',
-      display: 'Algorithms',
-      to: '/showcase/algorithms'
+      id: 'showcase-svg-map',
+      display: 'SVG Map',
+      to: '/showcase/svg-map'
     }]
   }],
   activeMenu: 'style-guide-alerts'